undo stocks changes
