+++ app/models/account.rb
@@ -6,7 +6,7 @@ class Account < ApplicationRecord
   has_many :balances, dependent: :destroy
   has_many :sender_transference, class_name: 'Transference', foreign_key: 'sender_id', dependent: :destroy
   has_many :receiver_transference, class_name: 'Transference', foreign_key: 'receiver_id', dependent: :destroy
-  has_many :treasuries, class_name: 'Investments::Treasury', dependent: :destroy
+  has_many :treasuries, class_name: 'Investments::Treasury::Treasury', dependent: :destroy
 
   monetize :balance_cents
