+++ spec/factories/dividends.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :dividend, class: 'Investments::Dividends' do
+  factory :dividend do
     stock
     value { Faker::Number.number(digits: 7) }
   end
